Moved AuditColumns into separate file
