+++ exasol/analytics/query_handler/query/select.py
@@ -4,6 +4,7 @@ from typing import (
     Optional,
 )
 
+from exasol.analytics.audit.columns import AuditColumns
 from exasol.analytics.query_handler.query.interface import Query
 from exasol.analytics.schema import (
     Column,
@@ -12,13 +13,10 @@ from exasol.analytics.schema import (
     DbOperationType,
 )
 
-DB_OBJECT_SCHEMA_TAG = "DB_OBJECT_SCHEMA"
-""" Optional, can be NULL """
-DB_OBJECT_NAME_TAG = "DB_OBJECT_NAME"
-""" Contains the schema name for operations CREATE/DROP SCHEMA """
-DB_OBJECT_TYPE_TAG = "DB_OBJECT_TYPE"
-DB_OPERATION_TYPE_TAG = "DB_OPERATION_TYPE"
-
+DB_OBJECT_SCHEMA_TAG = AuditColumns.OBJECT_SCHEMA.name.name
+DB_OBJECT_NAME_TAG = AuditColumns.OBJECT_NAME.name.name
+DB_OBJECT_TYPE_TAG = AuditColumns.OBJECT_TYPE.name.name
+DB_OPERATION_TYPE_TAG = AuditColumns.OPERATION_NAME.name.name
 
 class CustomQuery(Query):
     def __init__(self, query_string: str):
